ai-lens-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
Pillow==10.4.0
numpy==1.26.4
torch==2.4.1
torchvision==0.19.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
pydantic==2.9.2
scikit-learn==1.5.2
